Chore: fix incorrect configuration and move files
- component scan을 대비하여 api-user, api-admin, storage의 패키지 구조를 3 depth 고정으로 하기
- dependency에 api 문법을 사용하려면 java-library를 사용해야함
