--- api-admin/src/main/java/com/app/apiadmin/ApiAdminApplication.java
@@ -1,13 +1,0 @@
-package com.app.apiadmin;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ApiAdminApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ApiAdminApplication.class, args);
-    }
-
-}
