+++ api-admin/src/main/java/com/app/travelboard/ApiAdminApplication.java
@@ -1,13 +1,13 @@
-package com.app.apiuser;
+package com.app.travelboard;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class ApiUserApplication {
+public class ApiAdminApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(ApiUserApplication.class, args);
+        SpringApplication.run(ApiAdminApplication.class, args);
     }
 
 }
