+++ api-user/src/main/java/com/app/travelboard/ApiUserApplication.java
@@ -1,0 +1,15 @@
+package com.app.travelboard;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
+
+@ConfigurationPropertiesScan
+@SpringBootApplication
+public class ApiUserApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ApiUserApplication.class, args);
+    }
+
+}
