+++ api-user/src/test/java/com/app/travelboard/ApiUserApplicationTests.java
@@ -1,10 +1,10 @@
-package com.app.apiadmin;
+package com.app.travelboard;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class ApiAdminApplicationTests {
+class ApiUserApplicationTests {
 
     @Test
     void contextLoads() {
