+++ storage/build.gradle.kts
@@ -1,7 +1,5 @@
 dependencies {
-    implementation(project(":support"))
-    implementation("org.springframework.boot:spring-boot-starter-data-jpa")
-
+    api("org.springframework.boot:spring-boot-starter-data-jpa")
     runtimeOnly("com.h2database:h2")
     runtimeOnly("com.mysql:mysql-connector-j")
 }
