+++ storage/src/main/resources/db-local.yml
@@ -13,7 +13,6 @@ spring:
       enabled: true
       path: /h2-console
 
-storage:
   datasource:
     driver-class-name: org.h2.Driver
     url: jdbc:h2:mem:core;MODE=MySQL;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE
